requirements.txt
python-pptx==0.6.21
deepagents==0.1.0
langchain==0.1.0
langchain-openai==0.0.5
pydantic==2.5.0
python-dotenv==1.0.0
tavily==0.3.1
